fix(resource): select current game version when open resource downloader
